% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importannotation.R
\name{annotation_import}
\alias{annotation_import}
\title{Import annotations from Excel (xlsx or xsl) or text (txt) files}
\usage{
annotation_import()
}
\value{
A data frame ,pRoannot class object, containing the description of the variables within your dataset
}
\description{
This function is used to import annotations (descriptions) of
the variables present in the dataset.
}
\details{
This function creates a dataframe containing a short description of the variables
which can be further used in pRocessomics univariate
and multivariate statistical analyses.
These descriptions can come from Mapman categories or another source.
For further details about Mapman categories see \url{https://mapman.gabipd.org/app/mercator}
For Custom annotation files, there is a main required structure.
Annotation tables should have four columns: variable identifier or
accession, as named in the datasets; variable description,
as character; bin/group/pathway code, as numeric; and variable
bin/group/pathway name, as character.
Annotation tables must be in txt, xls, or xlsx format
This function will keep only the variables present in the datasets, due to this fact is
recommended to use this function after data pre-processing and filtering steps.
The information introduced will be available to include in graphics and tables.
Please go to \url{https://github.com/Valledor/pRocessomics/} for more information and examples.
}
\note{
Mapman annotation files could contain duplicates, importannotation function will ask you how to proceed.
}
\author{
Luis Valledor and Laura Lamelas
}
