% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bda_analysis.R
\name{bda_analysis}
\alias{bda_analysis}
\title{bda analysis}
\usage{
bda_analysis(datalist, omiclevel = NULL, annotation = NULL, 
initialrow = 1, initialcolumn = 3, treatment1col = 1, treatment2col = 2, 
treatment = 1, ncomponents = 2, keepX = NULL, folds = 3, nrepeat = 5,
autotune = FALSE, strength = 0.1)
}
\arguments{
\item{datalist}{A "POL" object, see preprocess_omic_list for further details}

\item{omiclevel}{Character vector containing the name/s of omics layers to be analyzed, if set to "all" note all
datasets will be concatenated and will conform the input data}

\item{annotation}{optional, annotations file class pRoAnnot}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{ncomponents}{Number of components to build uop the model}

\item{keepX}{Number of variables to take into account for the analysis, if set to NULL
all variables will be considered}

\item{folds}{Numeric, the folds in the Mfold cross-validation, maximum allowed 50}

\item{nrepeat}{Number of repeats to test the model}

\item{autotune}{logical, if set to TRUE several KeepX values will be tested and the better (less model error)
will be emplyed}

\item{strength}{Numeric, accuracy of design matrix}
}
\value{
bdaanalysis class object containing the bda analysis of input dataset
}
\description{
A function to perform bda (block discriminant analysis)
}
\author{
Luis Valledor and Laura Lamelas
}
