% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bda_analysis_wizard.R
\name{bda_analysis_wizard}
\alias{bda_analysis_wizard}
\title{block-DA wizard}
\usage{
bda_analysis_wizard("outputobjectname")
}
\arguments{
\item{outputobjectname}{Character, the quoted name of the result of the analysis}
}
\value{
The Block DA results, a "bdaanalysis" class object
}
\description{
A function for easily perform a principal component analysis
}
\author{
Luis Valledor and Laura Lamelas
}
