% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bda_plot.R
\name{bda_plot}
\alias{bda_plot}
\title{A function to plot DIABLO based analysis}
\usage{
bda_plot(bda_list, treatment = 1, plottype = "Distance",
variablespace = "xy", useannot = FALSE, fontsizes = c(14,10,16,12),
cutoff = 0.5, fortopscoring = c(1,20,"abs"), confidence = 0.9)
}
\arguments{
\item{bda_list}{diabloanalysis class object containing the diablo analysis of the data}

\item{treatment}{numeric value to determine the colors and symbols to be displayed in the plot, used in Score plot and Topscring plot. Available optiones are:
\itemize{
\item 1 to show different colors according to treatment 1 (treatment1col)
\item 2 to show different colors according to treatment 2 (treatment2col)
\item 3 to show different colors according to the combination of both treatments
\item 4 to show different colors according to treatment 1 and different symbols for treatment 2
}}

\item{plottype}{available options:
\itemize{
\item Distance; sample plot each dot represent a sample
\item Distance_Ellipse; distance plot with confidence ellipses by tretment
\item Topscoring; most representantive loadings according to Distance plot
\item Var; variable plot, each dot represent a variable. select dataset and annotation
\item Cim;
\item Network; network
\item Circos; circos plot
\item All; all of above
}}

\item{variablespace}{c("x","y","xy")}

\item{useannot}{logical, indicating if annotation should be used or not, for annotations use the splsda analysis should include it}

\item{fontsizes}{font sizes for plot elements}

\item{cutoff}{cutoff value for network prunning}

\item{fortopscoring}{vector indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}

\item{confidence}{numeric value between 0 and 1 for confidence ellipses in Distance_Ellipse plot}
}
\value{
plotly type plot
}
\description{
This function depicts the results obtained from a DIABLO analysis
}
\author{
Luis Valledor and Laura Lamelas
}
