% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_analysis_wizard.R
\name{da_analysis_wizard}
\alias{da_analysis_wizard}
\title{DA wizard}
\usage{
da_analysis_wizard("outputobjectname")
}
\arguments{
\item{outputobjectname}{Character, the quoted name of the result of the analysis}
}
\value{
DA results, a "daanalysis" class object
}
\description{
A function for easily perform single omics discriminantt analysis
}
\author{
Luis Valledor and Laura Lamelas
}
