% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_plot.R
\name{da_plot}
\alias{da_plot}
\title{A function to plot a Discriminant Analysis (DA)}
\usage{
da_plot(da_list, treatment = 1, plottype = "Distance", variablespace = "x",
useannot = FALSE, fontsizes = c(14,10,16,12), cutoff = 0.8, fortopscoring = c(1,20,"abs"), 
confidence = 0.9)
}
\arguments{
\item{da_list}{da analysis object, class "daanalysis"}

\item{treatment}{numeric value to determine the colors and symbols to be displayed in the plot, used in Score plot and Topscring plot. Available optiones are:
\itemize{
\item 1 to show different colors according to treatment 1 (treatment1col)
\item 2 to show different colors according to treatment 2 (treatment2col)
\item 3 to show different colors according to the combination of both treatments
\item 4 to show different colors according to treatment 1 and different symbols for treatment 2
}}

\item{plottype}{available options:
\itemize{
\item Distance, sample plot each dot represent a sample
\item Distance_Ellipse, distance plot with confidence ellipses by treatment
\item Topscoring, most representantive loadings according to Distance plot
\item Var, variable plot, each dot represent a variable. select dataset and annotation
\item Composite, combination of the two most representative plots
\item Composite_Annotated, same as composite but with annotations, recommended for exploratory analysis
\item Network, network
\item All all aforementioned plots will be exported as .pdf files in your working directory.
}}

\item{variablespace}{c("x","y","xy")}

\item{useannot}{logical, indicating if annotations should be used or not, please note the previous plsda analysis should include annotations}

\item{fontsizes}{font sizes for plot elements}

\item{cutoff}{cutoff value for network prunning}

\item{fortopscoring}{vector indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}

\item{confidence}{Numeric value between 0 and 1 for confidence ellipses in Distance_Ellipse plot}
}
\value{
plotly type plot
}
\description{
This function allows to fully depict the results from the Discriminant analysis
}
\author{
Luis Valledor and Laura Lamelas
}
