% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importfromexcel.R
\name{dataset_import}
\alias{dataset_import}
\title{Imports Omics datasets to R}
\usage{
dataset_import("datasetname")
}
\arguments{
\item{datasetname}{Character. The name to be assigned to the imported data, (optional)}
}
\value{
pRoDS class list containing all omics levels of the dataset and an auto-generated vector with the dataset structure.
}
\description{
This function imports datasets from Excel files and checks for the dataset structure
}
\details{
This function imports a spreadsheet from office or open office (xlsx or xls extensions) to R,
and creates a list to be further used in pRocessomics.
If there are more than one omics level within your dataset (from 1 to n, i.e. proteins, metabolites, transcripts),
they shall be integrated in the same spreadsheet file, one level per sheet. In this case, the sheet names will
be kept as the names of the imported R list.

\code{\link{dataset_import}} also performs some initial quality and consistency checks. If the imported data is not OK
it will give the user some hints for correcting it.
As a golden rule: cases (samples) data should be in rows, while variables should
be in columns (for large datasets i.e. transcriptomics data, it can be transposed during the importation process).
First column must indicate the name of the cases or samples, which
will be further treated as rownames in the rest of the functions of this package.
Sample names should be the same across the different omics levels. All omics levels
should have the same number of cases (this is a current limitation of this package).
We recommend that the second and third columns correspond to the different treatments
you applied to your experimental system, these factors must be also maintained all across the
different omics levels. And after these columns, those containing numeric variables (protein,
metabolite, RNA abundances, ...).
We strongly recommend that you employ unique protein, metabolite, gene, ... accessions
instead of definitions or descriptions for naming variables. Later you can upload
an annotation file to get the most of your data with \code{\link{annotation_import}} function.
Please go to \url{http://processomics.valledor.info} or \url{https://github.com/Valledor/pRocessomics/}
to download a template you can use to fill in with your data.
}
\seealso{
importannotation
}
\author{
Laura Lamelas and Luis Valledor
}
