% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{export_plot}
\alias{export_plot}
\title{A funtion for exporting pRocessomics generated plts to pdf format}
\usage{
export_plot(plot_object, filename="myplot.pdf", 
width=NULL, height=NULL, cutoff=NULL)
}
\arguments{
\item{plot_object}{name of the plot in the RStudio environment to be exported}

\item{filename}{quoted name of the exported plot with the extension .pdf}

\item{width}{width of the pdf object}

\item{height}{height of the pdf object}

\item{cutoff}{numeric, cutoff for the network, only used if the plot to be exported is a network}
}
\description{
This function exports plots to pdf format
}
\author{
Luis Valledor and Laura Lamelas
}
