% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R
\name{export_table}
\alias{export_table}
\title{Function for table exportation}
\usage{
export_table(result_list,filename=NULL)
}
\arguments{
\item{result_list}{name of the list to be exported in the local environment}

\item{filename}{quoted name to be given to the result_list object with the extension .xlsx}
}
\description{
This function exports pRocessomics generated tables to Excel format
}
\author{
Luis valledor and Laura Lamelas
}
