% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureselect.R
\name{featureselection}
\alias{featureselection}
\title{Feature Selection in Omic Datasets}
\usage{
featureselection(datalist, initialrow, initialcolumn, 
treatment1col, treatment2col, treatment, method = NULL, 
threshold = NULL, parametric = TRUE, stat = "p", abovebelow = "Below")
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. pRoDS class object.}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{method}{Method for selecting variables: "IQR"- interquartile range; "Stat" - ANOVA/Kruskal Wallis; "VarCoef" - Coeficient of variation.}

\item{threshold}{Desired threshold to apply to the data
For "IQR" : \% over the average IQR; For "Stat": p or q value; For "VarCoef" : Coeficient of variation}

\item{parametric}{boleean (TRUE, FALSE). If TRUE ANOVA test will be used, if FALSE Kruskal Wallis.}

\item{stat}{Indicates which statistic will be used p-values,"p", or q-values, "q". Only required when method = "Stat".}

\item{abovebelow}{character, set to "Above" for keep those variable- values above the threshold or "Below" to keep those variables below the threshold}
}
\value{
returns a "POL" object with the filtered data
}
\description{
This function processes a POL object to be analyzed in order to select featured variables
}
\details{
The objective of \code{featureselection} is removing those variables who do not greatly change between treatments, known to later introduce noise in multivariate analyses reducing the power of these approaches. This function requires a \code{POL} object, previously defined by using \code{preprocess_omic_list} and, if desired, \code{transformdata}. Two different approaches are implemented within this function: first one is based on selecting those variables that have intercuartilic ranges, IQR, greater than average within all studied variables. In this case the \code{threshold} indicates the \\% above average IQR and only variables with greater or equal IQR will be selected. The second strategy is based on applying ANOVA or Kruskal Wallis and selected those variables exhibiting a p or q value smaller or equal than \code{threshold}.
}
\author{
Luis Valledor and Laura Lamelas
}
