% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureselection_wizard.R
\name{featureselection_wizard}
\alias{featureselection_wizard}
\title{featureselection_wizard() easily filter more significant variablest}
\usage{
featureselection_wizard()
}
\value{
returns a "POL" object with the filtered data
}
\description{
This function is a wizard to interactively help the user to perform the feature selection on preprocessed data
}
\details{
The objective of \code{featureselection_wizard} is removing those variables who do (not)
greatly change between treatments,
known to later introduce noise in multivariate analyses reducing the power of these approaches.
This function requires a \code{POL} object, previously defined by using \code{preprocess_omic_list} and,
if desired, \code{transformdata}. Two different approaches are implemented within this
function: first one is based on selecting those variables that have inter-quartilic ranges,
IQR, greater than average within all studied variables. In this case the \code{threshold} i
ndicates the \\% above average IQR and only variables with greater or equal IQR will be selected.
The second strategy is based on applying ANOVA or Kruskal Wallis and selected those variables
exhibiting a p or q value smaller or equal than \code{threshold}.
}
\seealso{
featureselect

featureselection
}
\author{
Luis Valledor and Laura Lamelas
}
