% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica_analysis.R
\name{ica_analysis}
\alias{ica_analysis}
\title{A function to perform ica analysis}
\usage{
ica_analysis(datalist, annotation = NULL, initialrow = 1, 
initialcolumn = 3,  treatment1col = 1, treatment2col = 2, treatment,
ncomp = 3, omiclevel = "all")
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. POL class object.}

\item{annotation}{optional, pRoAnnot class object created with pRocessomics::importannotation function}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{ncomp}{Integer, number of components of the model}

\item{omiclevel}{Vector containing the name/s of omics layers to be analyzed}
}
\value{
icaanalysis class object
}
\description{
This function performs the Independent Component Analysis on omic datasets
}
\author{
Luis Valledor and Laura Lamelas
}
