% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica_analysis_wizard.R
\name{ica_analysis_wizard}
\alias{ica_analysis_wizard}
\title{ICA wizard}
\usage{
ica_analysis_wizard(outputobjectname)
}
\arguments{
\item{outputobjectname}{character, quoted name of the output of this analysis}
}
\value{
The ICA results, a "icaanalysis" class object
}
\description{
A function for easily perform a Independent Component Analysis
}
\author{
Luis Valledor and Laura Lamelas
}
