% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica_plot.R
\name{ica_plot}
\alias{ica_plot}
\title{A function to plot ica analysis}
\usage{
ica_plot (ica_list, treatment=1, compX=1, compY=2, 
compZ=NULL, plottype="Composite", useannot=FALSE, 
fontsizes=c(14,10,16,12), fortopscoring=c(1,20,"abs"))
}
\arguments{
\item{ica_list}{ica object to plot}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{compX}{component to be plot in X axis in Score and Var plots, default 1}

\item{compY}{component to be plot in X axis in Score and Var plots, default 2}

\item{compZ}{component to be plot in Z axis for 3D plots in Score and Var plots, default NULL}

\item{plottype}{available options:
Composite: combination of the two most representative plots (Score and Var)
Score, sample plot each dot represent a sample
Topscoring, top scoring variables in the selected (IC) independent component
Scree, explained variance plot
Var, variable plot, each dot represent a variable.
All all}

\item{useannot}{annotation file generated by pRocessomics::importannotation() function}

\item{fontsizes}{vector containing the font sizes to use in the plots}

\item{fortopscoring}{vector indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}
}
\value{
ica plot
}
\description{
This function makes different plots of ica analysis
}
\author{
Luis Valledor and Laura Lamelas
}
