% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importannotation.R
\name{importannotation}
\alias{importannotation}
\title{Import annotations from Excel (xlsx or xsl) or text (txt) files}
\usage{
importannotation()
}
\value{
A data frame ,pRoannot class object, containing the description of the variables within your dataset
}
\description{
This function is used to import annotations (descriptions) of
the variables present in the dataset.
}
\details{
This function creates a dataframe containing variables'
descriptions which can be further used in pRocessomics univariate
and multivariate statistical analysis.
These descriptions can come from Mapman categories or another source.
For Mapman annotation parser the returned txt can be used and this function
will keep only the variables present in the datasets, due to this fact is
recomended to use this function after data preprocessing and filtering steps.
For Custom annotation files, there is a main required sturcture.
Annotation tables should have four columns: variable identifier or
accesion, as named in the datasets; variable description,
as character; bin/group/pahtway code, as numeric; and variable
bin/group/pahtway name, as character.
The information introduced will be available to include in graphics and tables.
}
\note{
Mapman annotation files could contain duplicates, importannotation function will ask you how to proceed.
}
\author{
Luis Valledor and Laura Lamelas
}
