% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importfromexcel.R
\name{importfromexcel}
\alias{importfromexcel}
\title{Import datasets to R from an Excel file}
\usage{
importfromexcel("datasetname")
}
\arguments{
\item{datasetname}{Character. The name assigned to the imported data}
}
\value{
pRoDS class list containing all omic levels of the dataset and a vector with the dataset structure.
}
\description{
This function is used to import and check datasets from Excel files
}
\details{
This function imports a spreadsheet from office or open office (xlsx or xls extensions) to R,
creates a dataset list to be further used in pRocessomics.
It also performs some initial quality and consistency checks. If data is not OK
it will give the user some hints for correcting it.
The different omic levels within your dataset (from 1 to n) must be integrated
in the same Excel file, one sheet per level. Each Excel sheet should be named
with the name you want to give to each omic level.
As a golden rule: cases (samples) data should be in rows, while variables should
be in columns. First column must indicate the name of the cases or samples, which
will be further treated as rownames in the rest of the functions of this package.
Sample names should be the same across the different omic levels. All omic levels
should have the same number of cases (this is a current limitation of this package).
We recommend that second and third column will correspond to the different treatments
you applied to your data, these factors must be also maintained all across the
different omic levels. After these columns, variables (proteins,
metabolites, rnas, ...) should be indicated.
We strongly recommend that you employ unique protein, metabolite, gene, ... accessions
instead of definitions or descriptions for naming variables. Later you can upload
an annotation file to get the most of your data with \code{\link{importannotation}} function.
Please go to \url{http://processomics.valledor.info} or \url{https://github.com/Valledor/pRocessomics/}
to download a template you can use to fill with your data.
}
\seealso{
importannotation
}
\author{
Luis Valledor and Laura Lamelas
}
