% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_analysis.R
\name{kmeans_analysis}
\alias{kmeans_analysis}
\title{Kmeans analysis}
\usage{
kmeans_analysis(datalist,annotation=NULL,initialrow=1, initialcolumn=3, 
treatment1col=1, treatment2col=2, treatment=1,omiclevel="all",scalation=2,
clusters=c(10,20),show.elbow.plot=FALSE)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. POL class object.}

\item{annotation}{Character, quoted name of annotations object, an object generated by importannotation() function of class "annot"}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{omiclevel}{Character vector indicating the quoted name or names of the omiclevel(s) to be analyzed}

\item{scalation}{Numeric, indicating the type of scaling method to be applied to the data
Available scaling methods for k means analysis are:
scaling but not centering  <- 1
scaling and centering <- 2
column scaling expressed as decimal <- 3
row scaling expressed as decimal <- 4}

\item{clusters}{Vector (length 2) indicating the maximum and the minimum number of cluster to test}

\item{show.elbow.plot}{Boolean indicating whether elbow plot is desired to be displayed}
}
\value{
A "kmeananalysis" object, a list with the output of the kmeans analysis
}
\description{
A function to perform a k-means analysis
}
\author{
Luis Valledor and Laura Lamelas
}
