% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot.R
\name{kmeans_plot}
\alias{kmeans_plot}
\title{kmeans plot}
\usage{
kmeans_plot(kmeansobject, clusternumber=NULL, fontsizes=c(14,10,16,12))
}
\arguments{
\item{kmeansobject}{List containing kmeans analysis result generated with pRocessomics::kmeans_analysis()}

\item{clusternumber}{Numeric, Number of clusters to be represented. Note that the number of clusters has to be concordant with the respective analysis}

\item{fontsizes}{Vector (length 4) containing the font sizes to be used in the plot}
}
\value{
kmeans plot
}
\description{
A function to plot kmeans clustered groups
}
\author{
Luis Valledor and Laura Lamelas
}
