% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_wizard.R
\name{kmeans_wizard}
\alias{kmeans_wizard}
\title{kmeans wizard}
\usage{
kmeans_wizard()
}
\value{
kmeans result list and kmeans analysis plot
}
\description{
A function to ease the kmneans analysis
}
\author{
Luis Valledor and Laura Lamelas
}
