% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapman_wizard.R
\name{mapman_wizard}
\alias{mapman_wizard}
\title{mapman_wizard() easily analyze and depict mapman categories of your data}
\usage{
mapman_wizard(outputobjectname)
}
\arguments{
\item{outputobjectname}{character, name for mapman result object}
}
\value{
returns a "MMOG" object and its depiction
}
\description{
This function is a wizard to interactively help the user to perform the mapman analysis on preprocessed and annotated data
}
\seealso{
\code{\link{importannotation}}, \code{\link{mapman_group}}, \code{\link{mapman_plot}}
}
\author{
Luis Valledor and Laura Lamelas
}
