% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcia_analysis.R
\name{mcia_analysis}
\alias{mcia_analysis}
\title{Multiple coinertia analysis}
\usage{
mcia_analysis(datalist, annotation=NULL, initialrow=1, initialcolumn=3,treatment1col=1, 
treatment2col=2, treatment=1, omiclevel="all", keptaxes=5)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. POL class object.}

\item{annotation}{optional, pRoAnnot class object created with pRocessomics::importannotation
function containing the descriptions of the variables within the datasets}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{omiclevel}{Vector containing the name/s of omics layers to be analyzed}

\item{keptaxes}{Integer, numeber of omiclevels to be kept for the analysis}
}
\value{
A "mcoaanalysis" class list object containing the result of the mcia analysis
}
\description{
A function to perform a multiple coinertia analysis
}
\author{
Luis Valledor and Laura Lamelas
}
