% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcia_analysis_wizard.R
\name{mcia_analysis_wizard}
\alias{mcia_analysis_wizard}
\title{Multiple coinertia analysis}
\usage{
mcia_analysis_wizard(datalist=NULL, annotation=NULL, keptaxes=5, outputobjectname=NULL)
}
\arguments{
\item{datalist}{List with the different (at least two) preprocessed omic levels. POL class object.}

\item{annotation}{Character, Annotation file name.}

\item{keptaxes}{Integer, number of axes (omic levels) to keep}

\item{outputobjectname}{Name given to the analysis object}
}
\value{
multiple coinertia analysis object
}
\description{
A function to easily perform a multiple coinertia analysis on omic datasets
}
