% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcia_plot.R
\name{mcia_plot}
\alias{mcia_plot}
\title{Multiple coinertia plot}
\usage{
mcia_plot(mcoin_list,treatment=1, compX=1, 
compY=2, compZ=NULL, plottype="Composite",useannot=FALSE,
levelchoice="all",fontsizes=c(14,10,16,12),fortopscoring=c(1,20,"abs"))
}
\arguments{
\item{mcoin_list}{List containing the multiple coinertia analysis, generated with pRocessomics::mcia_analysis function}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{compX}{component to be plot in X axis in Score and Var plots, default 1}

\item{compY}{component to be plot in X axis in Score and Var plots, default 2}

\item{compZ}{component to be plot in Z axis for 3D plots in Score and Var plots, default NULL}

\item{plottype}{available options:
Composite: combination of the four most representative plots of the analysis
SynVar, sample plot each dot represent a sample
Topscoring, topscoring variables in the selected dimension
Scree, explained variance plot
Var, variable plot, each dot represent a variable.
CoIn, coinertia plot showing samples and omiclevels
Biplot, biplot showing samples and variables
Pseudoeigen, plot showing correlations among datasets
All all}

\item{useannot}{annotation file generated by pRocessomics::importannotation() function}

\item{levelchoice}{omiclevels (two or more) to analyze}

\item{fontsizes}{vector containing the font sizes to use in the plots}

\item{fortopscoring}{vector (length 3) indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}
}
\value{
mcia plot
}
\description{
A function to plot different aspects of multiple coinertia analysis
}
\author{
Luis Valledor and Laura Lamelas
}
