% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis.R
\name{pca_analysis}
\alias{pca_analysis}
\title{Principal Component Analysis}
\usage{
pca_analysis(datalist, annotation = NULL, initialrow = 1, initialcolumn = 3, 
treatment1col = 1, treatment2col = 2, treatment = 1, omiclevel = NULL)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. pRoDS class object.}

\item{annotation}{optional, pRoAnnot class object created with pRocessomics::importannotation
function containing the descriptions of the variables within the datasets}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{omiclevel}{Vector containing the name/s of omics layers to be analyzed}
}
\value{
A "pcaanalysis" class list containing the main outputs of the PCA which can be plotted by using pRocessomics::pca_plot() function
}
\description{
A function to perform a Principal Component Analysis on the given dataset or dataset levels
}
\details{
This function performs a PCA os scaled and centered data
}
\seealso{
pca_plot
}
\author{
Luis Valledor and Laura Lamelas
}
