% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_analysis_wizard.R
\name{pca_analysis_wizard}
\alias{pca_analysis_wizard}
\title{PCA wizard}
\usage{
pca_analysis_wizard()
}
\value{
The PCA results, a "pcaanalysis" class object
}
\description{
A function for easily perform a principal component analysis
}
\author{
Luis Valledor and Laura Lamelas
}
