% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA ploting function}
\usage{
pca_plot(pca_list,treatment=1, compX=1, compY=2, compZ=NULL, plottype="Composite",
useannot=FALSE,fontsizes=c(14,10,16,12),fortopscoring=c(1,20,"abs"))
}
\arguments{
\item{pca_list}{pca object to plot, created with pRocessomics::pca_analysis() function}

\item{treatment}{numeric value to determine the colors and symbols to be displayed in the plot, used in Score plot, Topscring plot and Biplot. Available optiones are:
\itemize{
\item 1 to show different colors according to treatment1,
\item 2 to show different colors according to treatment 2
\item 3 to show different colors according to the combination of both treatments
\item 4 to show different colors according to treatment 1 and different symbols for treatment 2
}}

\item{compX}{component to be plot in X axis in Score and Var plots, default 1}

\item{compY}{component to be plot in X axis in Score and Var plots, default 2}

\item{compZ}{component to be plot in Z axis for 3D plots in Score and Var plots, default NULL}

\item{plottype}{available options:
\itemize{
\item Scree, explained variance plot.
\item Score, sample plot each dot represent a sample.
\item Topscoring, top scoring variables in the selected (PC) principal component.
\item Biplot, PCA biplot showing samples and variables.
\item Var, variable plot, each dot represent a variable.
\item Composite: combination of the most representative plots (Score, Var and Biplot)
\item All all aforementioned plots will be exported as .pdf files in your working directory.
}}

\item{useannot}{annotation file generated by pRocessomics::importannotation() function}

\item{fontsizes}{vector containing the font sizes to use in the plots}

\item{fortopscoring}{vector indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}
}
\value{
pca plot
}
\description{
A function to plot Principal Component Analysis results
}
\author{
Luis Valledor and Laura Lamelas
}
