% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_omic_list.R
\name{preprocess_omic_list}
\alias{preprocess_omic_list}
\title{Preprocess datalist}
\usage{
preprocess_omic_list(datalist, initialrow=1, initialcolumn=2, 
treatment1col=1, treatment2col=2, treatment=1, imputation=NULL, imputhld=0.25, 
k=3, parallel=FALSE, varsel=FALSE, varselthld=0.25, abdbal=NULL)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. pRoDS class object.}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{imputation}{Method for imputation, "RF" - Random Forests,"KNN" - K-nearest neighbor, or "none". If NULL Random Forest methods will be employed.This can be provided as a single value or a vector with the same number of elements than list. Each vector element can be a different imputation method or none}

\item{imputhld}{Maximum number of NAs, in \%, per variable to be imputed.}

\item{k}{Number of neighbors employed in KNN algorithm.}

\item{parallel}{boleean. If parallel=TRUE missing value imputation algorithm will be run in parallel.}

\item{varsel}{boleean, If TRUE Variable selection based on consistency criteria will be used to pre-filter the data. Variables present in less than the stablished varselthld will be dropped out the analysis.}

\item{varselthld}{Minimum number of significant (not zero) values, in \%, per variable to be kept.}

\item{abdbal}{Abundance balancing normalization. "sample" - sample centric approach, "AvgIntensity" - data will be processed according to a sample centric approach and then each value is multipled by average intensity (sum of the intensities of all variables within a sample) of all samples. "TreatAvgIntensity" - data is sample centric normalized and then multiplied by the average intensity of samples of specific treatments. "none" - no abundance balancing is performed. If NULL "AvgIntensity" will be employed. This can be provided as a single value or a vector with the same number of elements than list. Each vector element can be a different abundance balancing method or none}
}
\value{
A POL class object, a list with the processed dataset
}
\description{
A function to preprocess omics data by imputing (or not) missing values, balance abundance among samples and pre-filter the data
}
\details{
The objective \code{preprocess_omic_list} is providing a first step of data preprocessing before uni and multivariate statistics are performed. The need of removing inconsistent variables (those present in only one or two samples, or really close to detection limit) and also defining which values are NA or zeroes for later imputation is a constant when working with omic datasets. To this end within this function the user can select the range of data to be analyzed, how it will be processed (considering all dataset together, or splitting it in the different treatments) towards the definition and imputation of missing values and also for balancing the abundance of each sample.
All tables must have the same structure with cases in rows and variables in columns. Row names, and columns in which treatments are defined should have the same name and order across the different datasets. By default, first columns should be devoted to indicate treatments. Row names should be unique, and preferable equal to database accession names in order to use later annotation steps (see example datasets).
First step of this function is deciding which value should be a 0 or NA. To this end all NAs are turned to 0, and then it is decided if specific values should remain as 0 or set as NA for imputation. \code{threshold} argument defines the maximum number of zeroes allowed for imputation within each variable and treatment (defined in \code{treatment}).
The second step of the function balance the abundance of the variables according to different methods: sample centric approach; average intensity of samples, or average intensity within treatment.
}
\author{
Luis Valledor and Laura Lamelas
}
