% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_wizard.R
\name{preprocess_wizard}
\alias{preprocess_wizard}
\title{Wizard for easily preprocess omic datasets}
\usage{
preprocess_wizard(outputobjectname)
}
\arguments{
\item{outputobjectname}{quoted name to assign to the preprocessed object}
}
\value{
A POL class list containing preprocessed omics data
}
\description{
This function is used to define the parameters and preprocess omic dataset lists in a friendly way.
}
\details{
This function pre-processes your data by handling missing values, differences in relative sample abundance and prefiltering. All the mentioned steps are optional and the user will be able to choose during the wizard run.
Please check \url{https://github.com/Valledor/pRocessomics/} for more information and examples.
}
\seealso{
\code{\link{preprocess_omic_list}}.
}
\author{
Luis Valledor and Laura Lamelas
}
