% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{select_palette}
\alias{select_palette}
\title{Select and store one of the pre-defined palettes for pRocessomics plots, pRocessomics functions will use the selected palettes to display the plots automatically if any pre-selected palette is in the global environment}
\usage{
select_palette()
}
\value{
charater vector, class "color.pRo" to set user defined colors/symbols/fontsizes to pRocessomics generated plots
}
\description{
This function allows to select de desired palette (among different options) to be displayed in pRocessomics plots.
}
\author{
Luis Valledor and Laura Lamelas
}
