% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletas.R
\name{set_custom_palettes}
\alias{set_custom_palettes}
\title{Set custom palette for pRocessomics plots}
\usage{
set_custom_palettes(palette_type,newcustomvalues)
}
\arguments{
\item{palette_type}{\itemize{
\item "mapman" For Mapman (or Custom annotation) related plots.
\item "levels" For different omic levels
\item "treatment" For different treatments
\item "continuous" For continuous scales
\item "symbols"
\item "fontsizes"
}}

\item{newcustomvalues}{character vector specifying desired values}
}
\value{
Override the default values to the user specified
}
\description{
This function allows to create a new palette to be displayed in pRocessomics plots. Including symbols, fontsizes and colors
}
\author{
Luis Valledor and Laura Lamelas
}
