% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spls_analysis_wizard.R
\name{spls_analysis_wizard}
\alias{spls_analysis_wizard}
\title{SPLS wizard}
\usage{
spls_analysis_wizard("outputobjectname")
}
\arguments{
\item{outputobjectname}{Character, the quoted name of the result of the analysis}
}
\value{
The SPL results, a "splsanalysis" class object
}
\description{
A function for easily perform a principal component analysis
}
\author{
Luis Valledor and Laura Lamelas
}
