% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spls_plot.R
\name{spls_plot}
\alias{spls_plot}
\title{sparse Partial Least Squares plot}
\usage{
spls_plot(spls_list,treatment=1, plottype="Distance",
variablespace="xy",useannot=FALSE,fontsizes=c(14,10,16,12),
cutoff=0.8,fortopscoring=c(1,20,"abs"))
}
\arguments{
\item{spls_list}{spls analysis object, class "splsanalysis"}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{plottype}{available options:
Composite; combination of the two most representative plots
Distance; sample plot each dot represent a sample
Topscoring; most representantive loadings according to Distance plot
Var; variable plot, each dot represent a variable. select dataset and annotation
Network; network
All; all of above}

\item{variablespace}{c("x","y","xy")}

\item{useannot}{logical, indicating if annotation should be used or not, for annotations use the splsda analysis should include it}

\item{fontsizes}{font sizes for plot elements}

\item{cutoff}{cutoff value for network prunning}

\item{fortopscoring}{vector indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}
}
\value{
plotly type plot
}
\description{
A function to perform already built spls analysis
}
\author{
Luis Valledor and Laura Lamelas
}
