% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_wizard.R
\name{transformation_wizard}
\alias{transformation_wizard}
\title{Omic data transformation Wizard}
\usage{
transformation_wizard()
}
\value{
Transformed data List
}
\description{
A fucntion to transform omic datasets
}
\details{
This wizard is aimed to guide you for properly transform your dataset. Please note that this module will only work if you have
preprocessed your data with preprocess_wizard or preprocess_omic_list pRocessomics functions.
}
\seealso{
importfromexcel, preprocess_omic_list, preprocess_wizard, transformdata
}
\author{
Luis Valledor and Laura Lamelas
}
