% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformdata.R
\name{transformdata()}
\alias{transformdata()}
\alias{transformdata}
\title{Data Transformation}
\usage{
transformdata(datalist, initialrow=1, initialcolumn=3, treatment1col = 1, 
treatment2col = 2, treatment = 3, transf=NULL)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. pRoDS class object.}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{transf}{Character vector (length = length datalist) determining the method used to transform each omiclevel. If only one value as transf="log" all the omiclevel will be log transformed, available options are:
z <- data scaling
z-centered <- data centering and scaling
log10 <- logaritmic transformation *please note negative values will yield a NaN
log10plus1 <- log(data+1)
box-cox-lambda <- box cox lambda transformation
sqrt <- square root *please note negative values are not allowed
cubic <- cubic root
arcsine <- arcsine *please note maximum absolute value is 100}
}
\value{
A list containing the transformed dataset
}
\description{
A function to transform omic data
}
\seealso{
transformation_wizard()
}
\author{
Luis Valledor and Laura Lamelas
}
