% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_plot.R
\name{univariate_plot}
\alias{univariate_plot}
\title{Univariate statistical analysis plot}
\usage{
univariate_plot(univarobject, plottype = "distribution", annotatefile = NULL, 
savetofile = F, filename = NULL)
}
\arguments{
\item{univarobject}{"UNIVAR" class object from univariate_analysis function}

\item{plottype}{\itemize{
\item distribution: scatter plot of p-values
\item histogram}}

\item{annotatefile}{optional, pRoAnnot class object created with pRocessomics::importannotation function}

\item{savetofile}{TRUE or FALSE, indicating if the plots should be}

\item{filename}{only applicable if savetofile set to TRUE}
}
\value{
p-value distribution plot or histogram
}
\description{
This function plots univariate statistical analyses by showing the p value distribution
}
\seealso{
univariate, univariate_wizard
}
\author{
Luis Valledor and Laura Lamelas
}
