% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_wizard.R
\name{univariate_wizard}
\alias{univariate_wizard}
\title{Univariate analysis wizard}
\usage{
univariate_wizard()
}
\description{
A function to perform univariate analysis within omic datasets
}
\details{
This wizard is aimed to guide you for properly analyze your datasets. Please note that this module will only work if you have preprocessed your data with preprocess_wizard or preprocess_omic_list functions.
During the wizard you will be asked to introduce your data, you will be allowed to check parametricity, transform your data and perform parametric and non-parametric tests to evaluate the significance of variables.
}
\seealso{
univariate, preprocess_omic_list, preprocess_wizard, transformadata, transformation_wizard
}
\author{
Luis Valledor and Laura Lamelas
}
