% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVdata.R
\docType{data}
\name{UVdataset}
\alias{UVdataset}
\title{Omic data}
\format{A list containing 4 data frames: proteome, metabolome, gene expression and physiology}
\usage{
data(UVdata)
}
\description{
Proteome, Metabolome, Gene expression & Physiological parameters of pine
needles exposed to UV radiation in a time-courses experiment.
}
