% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVdataset_ids.R
\docType{data}
\name{UVdataset_ids}
\alias{UVdataset_ids}
\title{Omic data annotations}
\format{A dataframe containing proteome and metabolome annotations}
\usage{
data(UVdata)
}
\description{
Proteome and Metabolome mercator mapman annotations
}
