% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venn_analysis.R
\name{Venn_analysis}
\alias{Venn_analysis}
\title{Venn analysis function}
\usage{
Venn_analysis(datalist,initialrow=1, initialcolumn=2, 
treatment1col=1,treatment2col=NA, treatment=1,omiclevel=NULL,threshold=2)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. pRoDS class object.}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{omiclevel}{Vector containing the name/s of omics layers to be analyzed}

\item{threshold}{Numeric indicating the minimum cases in which a variable has to be different to 0 to be taken into account}
}
\value{
A "Vennanalysis" class list object containing the Venn table
}
\description{
A function to sort variables in order of their presence or absence within the datasets
}
\author{
Luis Valledor and Laura Lamelas
}
