% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venn_diagram_wizard.R
\name{Venn_diagram_wizard}
\alias{Venn_diagram_wizard}
\title{Plot Venn diagrams}
\usage{
Venn_diagram_wizard()
}
\value{
The Venn diagram table and plot of an omic dataset
}
\description{
Plot Venn diagrams
}
\author{
Luis Valledor and Laura Lamelas
}
