% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venn_plot.R
\name{Venn_plot}
\alias{Venn_plot}
\title{Venn's Diagram plot}
\usage{
Venn_plot(datalist, Euler.dist = FALSE, alpha=0.5, num.cex=3,font.cex=3)
}
\arguments{
\item{datalist}{An object of class "vennanalysis" provided by Venn_analysis() function}

\item{Euler.dist}{Boolean indicating whether to draw Euler diagrams when conditions are met or not (Venn Diagrams with movable circles)}

\item{alpha}{Number giving the alpha transparency of circle areas}

\item{num.cex}{Number indicating the size of the numbers ploted}

\item{font.cex}{Number indicating the size of the font depicting the treatments/categories}
}
\value{
Venn Diagram plot
}
\description{
A function to plot the result of a Venn analysis
}
\details{
This function will automatically learn how many treatments or categories are to be depicted, in case of treatments > 4 a multi-layer circle plot will be provided
}
\author{
Luis Valledor and Laura Lamelas
}
