% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_analysis.R
\name{da_analysis}
\alias{da_analysis}
\title{A function to perform a sparce partial least squared discriminant analysis}
\usage{
da_analysis (datalist, annotation = NULL, initialrow=1, 
initialcolumn=3, treatment1col=1, treatment2col=2, treatment=1, 
omiclevel="all",keepX=NULL, autotune=FALSE, performance=FALSE, folds=3)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. POL class object.}

\item{annotation}{optional, pRoAnnot class object created with pRocessomics::importannotation
function containing the descriptions of the variables within the datasets}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{omiclevel}{Vector containing the name/s of omics layers to be analyzed}

\item{keepX}{Number of variables to take into account for the analysis, if set to NULL
all variables will be considered}

\item{autotune}{logical, if set to TRUE several KeepX values will be tested and the better (less model error)
will be employed}

\item{performance}{logical, analysis of the generated model performance in order to determine the
optimum number of caomponents}

\item{folds}{numeric, the folds in the Mfold cross-validation, maximum allowed 50}
}
\value{
da_analysis returns an object of class daanalysis, a list that contains the following components:
da: spls-da analysis output,
treatments: considered treatments in the analysis
datasetnames: the names of the input dataset
annotations: annotations (if provided) of the elements in the analysis
originaldata: original data
}
\description{
This function takes a preprocessed omic datalist and performs a spls-da, multivariate discriminant analysis
}
\details{
The aim of this function is to perform a supervised analysis by combining the assayed
experimental treatments against element values such as proteins/metabolites/transcripts
quantification profiles.
}
\seealso{
da_plot
}
\author{
Luis Valledor and Laura Lamelas
}
