% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica_plot.R
\name{ica_plot}
\alias{ica_plot}
\title{A function to plot Independent Components Analysis (ICA)}
\usage{
ica_plot (ica_list, treatment = 1, compX = 1, compY = 2,
compZ = NULL, plottype = "Composite", useannot = FALSE,
fontsizes = c(14,10,16,12), fortopscoring = c(1,20,"abs"),
confidence = 0.9)
}
\arguments{
\item{ica_list}{ica object to plot, class "icaanalysis"}

\item{treatment}{numeric value to determine the colors and symbols to be displayed in the plot, used in Score plot and Topscring plot. Available optiones are:
\itemize{
\item 1 to show different colors according to treatment 1 (treatment1col)
\item 2 to show different colors according to treatment 2 (treatment2col)
\item 3 to show different colors according to the combination of both treatments
\item 4 to show different colors according to treatment 1 and different symbols for treatment 2
}}

\item{compX}{component to be plotted in X axis in Score and Var plots, default 1}

\item{compY}{component to be plotted in X axis in Score and Var plots, default 2}

\item{compZ}{component to be plotted in Z axis for 3D plots in Score and Var plots, default NULL}

\item{plottype}{available options:
\itemize{
\item Scree, explained variance plot
\item Score, sample plot each dot represent a sample
\item Score_Ellipse, score plot with confidence ellipses by treatment
\item Topscoring, top scoring variables in the selected (IC) independent component
\item Var, variable plot, each dot represent a variable.
\item Composite: combination of the two most representative plots (Score and Var)
\item All all aforementioned plots will be exported as .pdf files in your working directory.
}}

\item{useannot}{logical, indicating if annotation should be used or not, please note the previous ica analysis should include it}

\item{fontsizes}{vector containing the font sizes to use in the plots}

\item{fortopscoring}{vector indicating the number of the component to be analyzed, the top n loadings to plot and the mode to compute "abs" for absolute value and "posneg" for positive and negative loading analyzed separately}

\item{confidence}{numeric value between 0 and 1 for confidence ellipses in Score_Ellipse plot}
}
\value{
ica plot
}
\description{
This function makes different plots of ica analysis
}
\author{
Luis Valledor and Laura Lamelas
}
