% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importfromexcel.R
\name{importfromexcel}
\alias{importfromexcel}
\title{Import datasets from an Excel file}
\usage{
importfromexcel("datasetname")
}
\arguments{
\item{datasetname}{The name assigned to the imported data}
}
\value{
List containing all omic levels of the dataset and a vector with the dataset structure.
}
\description{
This function is used to import and check datasets from Excel files
}
\details{
This function creates a dataset list to be further used in pRocessomics,
it also perform some initial quality and consistency checks. If data is not OK
it will give the user some hints to correct it.
The different omic levels within your dataset (from 1 to n) must be integrated
in the same Excel file, one sheet per level. Each Excel sheet should be named
with the name you want to give to each omic level.
As a golden rule: cases (samples) data will be in rows, while variables will
be in columns. First column must indicate the name of the cases or samples,
these names should be the same across the different omic levels. All omic levels
should have the same number of cases (this is a current limitation of this package).
We recommend that second and third column will correspond to the different treatments
you applied to your data, these factors must be also maintained all across the
different omic levels. After these columns, variables (proteins,
metabolites, rnas, ...) should be indicated.
We strongly recommend that you employ protein, metabolite, gene, ... accessions
instead of defline or description for naming variables. Later you can upload
an annotation file to get the most of your data (see importannotation()).
Please go to http://processomics.valledor.info or https://github.com/Valledor/pRocessomics/
to download a template you can use to fill your data.
}
\author{
Luis Valledor and Laura Lamelas
}
