% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot.R
\name{kmeans_plot}
\alias{kmeans_plot}
\title{kmeans plot}
\usage{
kmeans_plot(kmeansobject, clusternumber = NULL, treatment = 1,
plottype = "kmeansclust", collapsed = FALSE,  fontsizes = c(14,10,16,12))
}
\arguments{
\item{kmeansobject}{List containing kmeans analysis result generated with pRocessomics::kmeans_analysis()}

\item{clusternumber}{Numeric, Number of clusters to be represented. Note that the number of clusters has to be concordant with the respective kmeansobject
If set to NULL or not declared and selected plottype is kmeansclust all calculated plot will be displayed and in the case of heatmap plottype all variables will be displayed as heatmap rows.}

\item{treatment}{Numeric value to determine the treatments to be displayed in the x axes. Available options are:
\itemize{
\item 1 according to treatment 1 (treatment1col)
\item 2 according to treatment 2 (treatment2col)
\item 3 according to the combination of both treatments
\item 4 multiple-kmeans: X-axes according to treatment 2 and different lines for treatment 1
}}

\item{plottype}{Character, class of the plot to display available options are kmeansclust and heatmap}

\item{collapsed}{Boolean, TRUE indicates multiple overlap depiction}

\item{fontsizes}{Vector (length 4) containing the font sizes to be used in the plot}
}
\value{
kmeans plot
}
\description{
A function to plot kmeans clustered groups
}
\author{
Luis Valledor and Laura Lamelas
}
