% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapman_group.R
\name{mapman_group}
\alias{mapman_group}
\title{Mapman group}
\usage{
mapman_group(datalist, annotation, initialrow=1, initialcolumn=3, treatment1col=1, 
treatment2col=2, treatment=1, group=TRUE, omiclevel=NULL, 
threshold=c(0,0), stats=FALSE, scalevalues=TRUE, 
normbyvariables=FALSE)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. pRoDS class object.}

\item{annotation}{"annot" class object containing variables descriptions according to mercaror mapman, generated from pRocessomics::importannotation() function}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{group}{Boolean to indicating whether replicates from same treatment should be grouped. If set to positive average values will be shown}

\item{omiclevel}{Character vector indicating the quoted name or names of the omiclevel(s) to be analyzed}

\item{threshold}{Vector (length 2) indicating number of minimum replicates different to 0 belonging to a variable to consider that variable present for that treatment, and number of minimum variables in a Mapman category to consider that category. Default c(0,0)}

\item{stats}{Boolean indicating whether stats calculations as standard deviation and pvalues are returned with the mapman table}

\item{scalevalues}{Boolean indicating whether variables values should be scaled}

\item{normbyvariables}{Boolean indicating whether scale should be done according to variables (TRUE) or samples (FALSE)}
}
\value{
Mapman aggrupation table with sums of abundances according to mapman categories
}
\description{
A function to sort and aggrupate the variables in Mapman categories
}
\author{
Luis Valledor and Laura Lamelas
}
