% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapman_plot.R
\name{mapman_plot}
\alias{mapman_plot}
\title{Generate plots based on Mapman categories and groups (mapman_group ())}
\usage{
mapman_plot(MMOlist, plottype="bar", ploterror=TRUE, logscale=FALSE, 
hmparameters=c("manhattan","manhattan","ward.D",FALSE,"row"),normbyelements=FALSE,
fontsizes=c(14,10,16,12))
}
\arguments{
\item{MMOlist}{Mapman aggrupation table with sums of abundances according to mapman categories}

\item{plottype}{Character.Type of plot to be displayed, "bar", "heatmap" or "pie".}

\item{ploterror}{Boolean indicating whether errors within replicates should be shown}

\item{logscale}{Boolean indicating whether data should be log scaled}

\item{hmparameters}{Character vector (length 5) determining main features of heatmap plot if selected in plottype.
hmparameters <- c("distance measure used in row clustering","distance measure in column clustering","clustering method","displaynumbers","scale")
where distance measure used in row/col clustering: character, those methods supported by stats::dist ("correlation","euclidean","maximum","manhattan","camberra","binary","minkowski")
clustering method: character, those methods provided by stats::hclust ("ward.D","ward.D2","single","complete","average","mcquiry","median","centroid")
displaynumbers: boolean determining if numeric values should be indicated in heatmap cells.
scale: character, direction of scaling and centering ("row", "column","none")}

\item{normbyelements}{Boolean indicating if elements should be scaled}

\item{fontsizes}{fontsizes of title, subtitle and axis labels}
}
\value{
A data frame ,pRoannot class object, containing the description of the variables within your dataset
}
\description{
A function which is used to plot abundances of mapman_group object. Barplot, pie charts and heatmaps available
}
\details{
This function creates a dataframe containing variables'
descriptions which can be further used in pRocessomics univariate
and multivariate statistical analysis.
These descriptions can come from Mapman categories or another source.
For Mapman annotation parser the returned txt can be used and this function
will keep only the variables present in the datasets, due to this fact is
recomended to use this function after data preprocessing and filtering steps.
For Custom annotation files, there is a main required sturcture.
Annotation tables should have four columns: variable identifier or
accesion, as named in the datasets; variable description,
as character; bin/group/pahtway code, as numeric; and variable
bin/group/pahtway name, as character.
The information introduced will be available to include in graphics and tables.
}
\note{
Mapman annotation files could contain duplicates, importannotation function will ask you how to proceed.
}
\author{
Luis Valledor and Laura Lamelas
}
