% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcia_analysis_wizard.R
\name{mcia_analysis_wizard}
\alias{mcia_analysis_wizard}
\title{Multiple coinertia analysis}
\usage{
mcia_analysis_wizard(outputobjectname = NULL)
}
\arguments{
\item{outputobjectname}{Name given to the analysis object}
}
\value{
multiple coinertia analysis object
}
\description{
A function to easily perform a multiple coinertia analysis on omic datasets
}
