% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRocessomics.R
\docType{package}
\name{pRocessomics-package}
\alias{pRocessomics}
\alias{pRocessomics-package}
\title{pRocessomics: Omics Data Processing, Integration and Visualization Toolkit}
\description{
User friendly package for single and multi omics data analysis and visualization.
}
\details{
This is pRocessomics R package devoted to preprocess, filter and analyse omics data
}
\keyword{internal}
