% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_wizard.R
\name{transformation_wizard}
\alias{transformation_wizard}
\title{transformation_wizard(): easily transform omic datasets}
\usage{
transformation_wizard()
}
\value{
Transformed data List
}
\description{
A function for quickly and easily transformation of omic datasets
}
\details{
This wizard is aimed to guide you for properly transform your dataset. Please note that this module will only work if you have
preprocessed your data with preprocess_wizard() or preprocess_omic_list() functions.
}
\seealso{
\code{\link{dataset_import}}, \code{\link{preprocess_omic_list}}, \code{\link{preprocess_wizard}}, \code{\link{transformdata}}.
}
\author{
Luis Valledor and Laura Lamelas
}
