% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Univariate statistical analysis.}
\usage{
univariate(datalist,initialrow=1,initialcolumn=3, treatment1col=1,
treatment2col=2, treatment=1, parametric=TRUE, posthoc=TRUE, FDR=TRUE, 
round=6, annotatefile=NULL)
}
\arguments{
\item{datalist}{List with different preprocessed omic levels. POL class object.}

\item{initialrow}{First row of numeric data within datasets.}

\item{initialcolumn}{First column of numeric data within datasets.}

\item{treatment1col}{Column number in which treatment 1 is indicated.}

\item{treatment2col}{Column number in which treatment 2 is indicated (if applicable).
If there is only one treatment indicate the same column as treatment1col.}

\item{treatment}{numeric. Set 1 to split the dataset according to treatment1, 2 to split the
dataset according to treatment 2 and 3 to split the dataset according to the combination of both treatments.}

\item{parametric}{logical, this argument indicates whether the employed test
should be or not parametric, depending on the number of available treatments if parametric a t test or ANOVA
test will be employed, for non parametric data, Wilcox test or kruskal Wallis test will be employed.}

\item{posthoc}{logical, this argument indicates whether to do a posthoc test, if parametric=TRUE a
TukeyHSD test will be done, for non-parametric analysis Dunn test.}

\item{FDR}{logical, if TRUE adjusted pvalues will be provided according to Bonferroni method}

\item{round}{numeric, number of desired decimal digits in the output}

\item{annotatefile}{optional, pRoAnnot class object created with pRocessomics::importannotation function}
}
\value{
List containing mean, standard deviation and selected stats from each omic layer,
UNIVAR class object.
}
\description{
This function performs univariate statistical analyses such as
parametrics as t test or ANOVA and non-parametric tests as Wilcox test and
Kruskal Wallis, there is also the posibility to adjust obtained pvalues,
FDR and post hoc analyses.
}
\details{
The objective of univariate function is to perform statistical analyses
in order to check variability across treatments.
This function requires a POL object, previously defined by using
preprocess_omic_list or preprocess_wizard and, if desired, transform_varfilter_list.
Two different approaches are implemented within this function, the first related
to parametriciy, and the second taking into account the number of treatments to
compare.
}
\seealso{
transformdata and featureselect
}
\author{
Luis Valledor and Laura Lamelas
}
